Consolidate cleaning implementation and tests
